// src/types/FinData.ts
export type FinData = {
    id?: number
    segmentId?: number
    quarter?: number
    year?: number
    netIncome?: number
    totalAssets?: number
    totalEquity?: number
    totalLiabilities?: number
    salesRevenue?: number
    marketValue?: number
    cashFlow?: number
    segmentName?: string
    companyName?: string
};